+++ src/core/components/profile/registercard.component.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Card, Container, Typography, Button, makeStyles } from "@material-ui/core"
-import { ProfileStatus } from "../../models/statusInfo.model"
+import { ProfileStatus } from "../../models/profileStatus.model"
 import Gear from "../../../assets/images/icon/gear-icon.svg"
 import { Link } from "react-router-dom"
